Removing old battle system
