+++ src/NewBattle.h
@@ -2,7 +2,9 @@
 	
 typedef struct BattleActor BattleActor;
 
-bool ClosingWhileInNewBattle(void);
+bool ClosingWhileInBattle(void);
+int GetCurrentMonsterHealth(void);
+void SaveBattleState(void);
 
 void ResumeBattle(int currentMonster, int currentMonsterHealth);
 bool IsBattleForced(void);
@@ -12,3 +14,6 @@ void UpdateNewBattle(void);
 BattleActor *GetPlayerActor(void);
 BattleActor *GetMonsterActor(void);
 void TriggerBattleScreen(void);
+void ResumeBattle(int currentMonster, int currentMonsterHealth);
+bool IsBattleForced(void);
+bool ClosingWhileInBattle(void);
